backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
python-dotenv==1.0.1
google-generativeai==0.5.4
httpx==0.27.0
Pillow==10.3.0
python-multipart==0.0.9
